requirements-minimal.txt
# Core dependencies
python-dotenv==1.0.1
fastapi==0.104.1
uvicorn==0.24.0
torch==2.2.0
asyncio>=3.4.3

# Base LangChain
langchain>=0.1.2
langchain-core>=0.1.45
langchain-community>=0.0.28

# Ollama integration
langchain-ollama>=0.1.0
